fix: replace .float() with .to(tensor.dtype) throughout forward path
All modules were hardcoding float32 via .float() on masks and one-hot
tensors, which breaks when the model runs in bfloat16 (the mask gets
multiplied or added to a bfloat16 tensor causing dtype mismatch errors).

Immediate crash: thalamus top_mask.float() → LayerNorm "expected Float
but found BFloat16". Fixed across all affected modules:
  - thalamus.py:  top_mask → .to(probs.dtype)
  - world_model.py: one_hot → .to(probs.dtype)
  - motor.py:     survival mask → .to(logits.dtype)
  - causal_inference.py: do_mask → .to(A.dtype)
  - cerebellum.py: threshold mask → .to(sparse.dtype)
  - vesicles.py:  active/new_pos/dead masks → .to(buffer.dtype)

+++ neuroslm/modules/causal_inference.py
@@ -158,7 +158,7 @@ class CausalInferenceModule(nn.Module):
         # Cut incoming edges at intervened nodes
         if do_mask is not None:
             # Zero column i of A for each intervened token i
-            A_mod = A * (1 - do_mask.float().unsqueeze(1))
+            A_mod = A * (1 - do_mask.to(A.dtype).unsqueeze(1))
         else:
             A_mod = A
